web: refactor locale related code
Abstract away localization code to a struct containing
a map of localizers and matchers. This new struct includes
a factory function and a method which replaces the LocalizerFunc
closure creator's job. This struct is included as a parameter
to newTemplateCache - renamed from initTemplateCache - and has
it's method used on the template funcmap.